Move SMIR projections tests to new file
